Camera2: add extended volume key and power key control
This commit introduces several things for camera/video/panorama mode:

- volume key mode control (VolumeUp & VolumeDown):
  - shutter/shutter with pre focus
  - focus/shutter (for video and panorama only shutter)
  - shutter/ focus (for video and panorama only shutter)
  - zoom up / zoom down (default) (only camera and video)

- power key control (enabled by default):
  - on/off power shutter with pre focus

As well this commit fixes the updatepreferences which is not correct called
on camera change. All possible updates are put into the handler que now.

Moved initTruView and initSmartCapture into main preference
methods.

Changed app to a priv-app to avoid to sign it with platform certificate so that
user can use this app as well on other roms with different certificate

Implemented method to prevend any key action when not in camera application
(like filmstripview or when menu is up etc)

several small fixes and cleanup

Thanks to CAF for the basic zoom code

Patchset:
- add zoom indicator to zoom feature
- try catch for power key override if a user try to install it on
  a rom without override support. Disables it completly.
- small cleanup

Patchet:
- fix sharedpreferences derp

PatchSet:
- add more zoomsteps for smoother zooming

Change-Id: Icc2e17cfde019c4d87202c0daea6eec380f3a204